requirements.txt
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-community>=0.0.10
faiss-cpu==1.7.4
chromadb==0.4.18
openai>=1.3.7
python-dotenv==1.0.0
streamlit>=1.28.1
pandas>=2.1.3
numpy<2.0.0,>=1.24.3
tiktoken>=0.5.1
sentence-transformers>=2.2.2
pypdf2>=3.0.1
python-docx>=1.1.0
beautifulsoup4>=4.12.2
requests>=2.31.0
pydantic-settings>=2.0.3
